rename to pypi-artifacts and combine into one package
